requirements.txt
seleniumbase==4.12.8
pytube==12.1.2
webdriver_manager==3.8.5
streamlit==1.18.0